fix：完善拼团活动CRUD

+++ yudao-module-mall/yudao-module-promotion-biz/src/main/java/cn/iocoder/yudao/module/promotion/service/combination/CombinationActivityServiceImpl.java
@@ -1,9 +1,11 @@
 package cn.iocoder.yudao.module.promotion.service.combination;
 
 import cn.hutool.core.collection.CollUtil;
+import cn.hutool.core.collection.CollectionUtil;
 import cn.hutool.core.util.ObjectUtil;
 import cn.iocoder.yudao.framework.common.enums.CommonStatusEnum;
 import cn.iocoder.yudao.framework.common.pojo.PageResult;
+import cn.iocoder.yudao.framework.common.util.collection.CollectionUtils;
 import cn.iocoder.yudao.module.product.api.sku.ProductSkuApi;
 import cn.iocoder.yudao.module.product.api.sku.dto.ProductSkuRespDTO;
 import cn.iocoder.yudao.module.product.api.spu.ProductSpuApi;
@@ -13,6 +15,7 @@ import cn.iocoder.yudao.module.promotion.controller.admin.combination.vo.activit
 import cn.iocoder.yudao.module.promotion.controller.admin.combination.vo.activity.CombinationActivityPageReqVO;
 import cn.iocoder.yudao.module.promotion.controller.admin.combination.vo.activity.CombinationActivityUpdateReqVO;
 import cn.iocoder.yudao.module.promotion.controller.admin.combination.vo.product.CombinationProductCreateReqVO;
+import cn.iocoder.yudao.module.promotion.controller.admin.combination.vo.product.CombinationProductUpdateReqVO;
 import cn.iocoder.yudao.module.promotion.convert.combination.CombinationActivityConvert;
 import cn.iocoder.yudao.module.promotion.dal.dataobject.combination.combinationactivity.CombinationActivityDO;
 import cn.iocoder.yudao.module.promotion.dal.dataobject.combination.combinationactivity.CombinationProductDO;
@@ -22,15 +25,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 
 import javax.annotation.Resource;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Set;
 
 import static cn.iocoder.yudao.framework.common.exception.util.ServiceExceptionUtil.exception;
 import static cn.iocoder.yudao.module.product.enums.ErrorCodeConstants.SPU_NOT_EXISTS;
-import static cn.iocoder.yudao.module.promotion.enums.ErrorCodeConstants.COMBINATION_ACTIVITY_NOT_EXISTS;
-import static cn.iocoder.yudao.module.promotion.enums.ErrorCodeConstants.COMBINATION_ACTIVITY_SPU_CONFLICTS;
+import static cn.iocoder.yudao.module.promotion.enums.ErrorCodeConstants.*;
 import static cn.iocoder.yudao.module.promotion.util.PromotionUtils.validateProductSkuExistence;
 
 /**
@@ -54,9 +55,9 @@ public class CombinationActivityServiceImpl implements CombinationActivityServic
     @Override
     public Long createCombinationActivity(CombinationActivityCreateReqVO createReqVO) {
         // 校验商品 SPU 是否存在是否参加的别的活动
-        validateProductCombinationConflict(createReqVO.getSpuIds());
+        validateProductCombinationConflict(createReqVO.getSpuId(), null);
         // 获取所选 spu下的所有 sku
-        List<ProductSkuRespDTO> skus = productSkuApi.getSkuListBySpuId(createReqVO.getSpuIds());
+        List<ProductSkuRespDTO> skus = productSkuApi.getSkuListBySpuId(CollectionUtil.newArrayList(createReqVO.getSpuId()));
         // 校验商品 sku 是否存在
         validateProductSkuExistence(skus, createReqVO.getProducts(), CombinationProductCreateReqVO::getSkuId);
 
@@ -79,21 +80,20 @@ public class CombinationActivityServiceImpl implements CombinationActivityServic
         return activityDO.getId();
     }
 
-    private void validateProductCombinationConflict(List<Long> spuIds) {
+    private void validateProductCombinationConflict(Long spuId, Long activityId) {
         // 校验商品 spu 是否存在
-        List<ProductSpuRespDTO> spuList = productSpuApi.getSpuList(spuIds);
-        if (ObjectUtil.notEqual(spuIds.size(), spuList.size())) {
+        List<ProductSpuRespDTO> spuList = productSpuApi.getSpuList(CollUtil.newArrayList(spuId));
+        if (CollUtil.isEmpty(spuList)) {
             throw exception(SPU_NOT_EXISTS);
         }
-        // 查询所有开启的秒杀活动
+        // 查询所有开启的拼团活动
         List<CombinationActivityDO> activityDOs = combinationActivityMapper.selectListByStatus(CommonStatusEnum.ENABLE.getStatus());
+        // 更新时排除自己
+        if (activityId != null) {
+            activityDOs.removeIf(item -> ObjectUtil.equal(item.getId(), activityId));
+        }
         // 过滤出所有 spuIds 有交集的活动
-        List<CombinationActivityDO> doList = activityDOs.stream().filter(s -> {
-            // 判断 spu 是否有交集
-            ArrayList<Long> spuIdsClone = CollUtil.newArrayList(s.getSpuIds());
-            spuIdsClone.retainAll(spuIds);
-            return CollUtil.isNotEmpty(spuIdsClone);
-        }).collect(Collectors.toList());
+        List<CombinationActivityDO> doList = CollectionUtils.convertList(activityDOs, c -> c, s -> ObjectUtil.equal(s.getId(), spuId));
         if (CollUtil.isNotEmpty(doList)) {
             throw exception(COMBINATION_ACTIVITY_SPU_CONFLICTS);
         }
@@ -102,29 +102,83 @@ public class CombinationActivityServiceImpl implements CombinationActivityServic
     @Override
     public void updateCombinationActivity(CombinationActivityUpdateReqVO updateReqVO) {
         // 校验存在
-        validateCombinationActivityExists(updateReqVO.getId());
+        CombinationActivityDO activityDO = validateCombinationActivityExists(updateReqVO.getId());
+        // 校验状态
+        if (ObjectUtil.equal(activityDO.getStatus(), CommonStatusEnum.DISABLE.getStatus())) {
+            throw exception(COMBINATION_ACTIVITY_STATUS_DISABLE);
+        }
+        // 校验商品冲突
+        validateProductCombinationConflict(updateReqVO.getSpuId(), updateReqVO.getId());
+        // 获取所选 spu下的所有 sku
+        List<ProductSkuRespDTO> skus = productSkuApi.getSkuListBySpuId(CollectionUtil.newArrayList(updateReqVO.getSpuId()));
+        // 校验商品 sku 是否存在
+        validateProductSkuExistence(skus, updateReqVO.getProducts(), CombinationProductUpdateReqVO::getSkuId);
+
         // 更新
         CombinationActivityDO updateObj = CombinationActivityConvert.INSTANCE.convert(updateReqVO);
         combinationActivityMapper.updateById(updateObj);
+        // 更新商品
+        updateCombinationProduct(updateObj, updateReqVO.getProducts());
+    }
+
+    /**
+     * 更新秒杀商品
+     *  TODO 更新商品要不要封装成通用方法
+     *
+     * @param updateObj DO
+     * @param products  商品配置
+     */
+    private void updateCombinationProduct(CombinationActivityDO updateObj, List<CombinationProductUpdateReqVO> products) {
+        List<CombinationProductDO> combinationProductDOs = combinationProductMapper.selectListByActivityId(updateObj.getId());
+        // 数据库中的活动商品
+        Set<Long> convertSet = CollectionUtils.convertSet(combinationProductDOs, CombinationProductDO::getSkuId);
+        // 前端传过来的活动商品
+        Set<Long> convertSet1 = CollectionUtils.convertSet(products, CombinationProductUpdateReqVO::getSkuId);
+        // 删除后台存在的前端不存在的商品
+        List<Long> d = CollectionUtils.filterList(convertSet, item -> !convertSet1.contains(item));
+        if (CollUtil.isNotEmpty(d)) {
+            combinationProductMapper.deleteBatchIds(d);
+        }
+        // 前端存在的后端不存在的商品
+        List<Long> c = CollectionUtils.filterList(convertSet1, item -> !convertSet.contains(item));
+        if (CollUtil.isNotEmpty(c)) {
+            List<CombinationProductUpdateReqVO> vos = CollectionUtils.filterList(products, item -> c.contains(item.getSkuId()));
+            List<CombinationProductDO> productDOs = CombinationActivityConvert.INSTANCE.convertList(updateObj, vos);
+            combinationProductMapper.insertBatch(productDOs);
+        }
+        // 更新已存在的商品
+        List<Long> u = CollectionUtils.filterList(convertSet1, convertSet::contains);
+        if (CollUtil.isNotEmpty(u)) {
+            List<CombinationProductUpdateReqVO> vos = CollectionUtils.filterList(products, item -> u.contains(item.getSkuId()));
+            List<CombinationProductDO> productDOs = CombinationActivityConvert.INSTANCE.convertList1(updateObj, vos, combinationProductDOs);
+            combinationProductMapper.updateBatch(productDOs);
+        }
     }
 
     @Override
     public void deleteCombinationActivity(Long id) {
         // 校验存在
-        validateCombinationActivityExists(id);
+        CombinationActivityDO activityDO = validateCombinationActivityExists(id);
+        // 校验状态
+        if (ObjectUtil.equal(activityDO.getStatus(), CommonStatusEnum.ENABLE.getStatus())) {
+            throw exception(COMBINATION_ACTIVITY_DELETE_FAIL_STATUS_NOT_CLOSED_OR_END);
+        }
+
         // 删除
         combinationActivityMapper.deleteById(id);
     }
 
-    private void validateCombinationActivityExists(Long id) {
-        if (combinationActivityMapper.selectById(id) == null) {
+    private CombinationActivityDO validateCombinationActivityExists(Long id) {
+        CombinationActivityDO activityDO = combinationActivityMapper.selectById(id);
+        if (activityDO == null) {
             throw exception(COMBINATION_ACTIVITY_NOT_EXISTS);
         }
+        return activityDO;
     }
 
     @Override
     public CombinationActivityDO getCombinationActivity(Long id) {
-        return combinationActivityMapper.selectById(id);
+        return validateCombinationActivityExists(id);
     }
 
     @Override
@@ -140,6 +194,11 @@ public class CombinationActivityServiceImpl implements CombinationActivityServic
     @Override
     public List<CombinationActivityDO> getCombinationActivityList(CombinationActivityExportReqVO exportReqVO) {
         return combinationActivityMapper.selectList(exportReqVO);
+    }
+
+    @Override
+    public List<CombinationProductDO> getProductsByActivityId(Long id) {
+        return combinationProductMapper.selectListByActivityId(id);
     }
 
 }
